Add notebook that does LDA in Gensim on the folia data
So far, in the notebook a dictionary of nouns is created (and
saved to disk), and a bow representation of the corpus is created (and
saved to disk). Next step is to do the LDA. Creating the dict and bow
representation is rather slow (~30 minutes each for ~1400 documents).
Perhaps it is faster to create text files containing the nouns. Also,
the number of nouns per document varies wildly (ranging from 1 to
thousands).